moving header components to header file
