+++ applications/jupyter-extension/nteract_on_jupyter/app/contents/index.tsx
@@ -4,20 +4,13 @@
 import * as React from "react";
 import { hot } from "react-hot-loader";
 import { connect } from "react-redux";
-import { Dispatch } from "redux";
 import { dirname } from "path";
 import * as actions from "@nteract/actions";
-import { NotebookMenu } from "@nteract/connected-components";
 import { AppState, ContentRef, selectors } from "@nteract/core";
-import { LoadingIcon, SavingIcon, ErrorIcon } from "@nteract/iron-icons";
-import { H4 } from "@blueprintjs/core";
 import urljoin from "url-join";
 
 import { ConnectedDirectory } from "./directory";
-import { EditableTitleOverlay } from "../components/editable-title-overlay";
-import { ThemedLogo } from "../components/themed-logo";
-import { Nav, NavSection } from "../components/nav";
-import LastSaved from "../components/last-saved";
+import { DirectoryHeader, ConnectedFileHeader as FileHeader } from "./headers";
 import { default as File } from "./file";
 
 interface IContentsProps {
@@ -37,108 +30,6 @@ interface IContentsProps {
 
 interface IContentsState {
   isDialogOpen: boolean;
-}
-
-const DirectoryHeader = (props: Partial<IContentsProps>) => (
-  <Nav>
-    <NavSection>
-      {props.appBase ? (
-        <a href={urljoin(props.appBase)} role="button" title="Home">
-          <ThemedLogo />
-        </a>
-      ) : null}
-    </NavSection>
-  </Nav>
-);
-
-class FileHeader extends React.PureComponent<
-  Partial<IContentsProps>,
-  IContentsState
-> {
-  constructor(props: Partial<IContentsProps>) {
-    super(props);
-
-    this.state = {
-      isDialogOpen: false
-    };
-  }
-
-  // Determine the file handler
-  getFileHandlerIcon = () => {
-    return this.props.saving ? (
-      <SavingIcon />
-    ) : this.props.error ? (
-      <ErrorIcon />
-    ) : this.props.loading ? (
-      <LoadingIcon />
-    ) : (
-      ""
-    );
-  };
-
-  // TODO: Add more acceptable file extensions
-  hasFileExtension = (fileName: string) => {
-    if (/\.ipynb/.exec(fileName) !== null) {
-      return true;
-    } else {
-      return false;
-    }
-  };
-
-  // TODO: Add logic for acceptable file extensions
-  addFileExtension = (fileName: string) => {
-    if (!this.hasFileExtension(fileName)) {
-      return `${fileName}.ipynb`;
-    } else {
-      return fileName;
-    }
-  };
-
-  openDialog = () => this.setState({ isDialogOpen: true });
-  closeDialog = () => this.setState({ isDialogOpen: false });
-
-  // Handles onConfirm callback for EditableText component
-  confirmTitle = (value: string) => {
-    if (value !== this.props.displayName) {
-      this.props.changeContentName({
-        contentRef: this.props.contentRef,
-        filepath: `/${value ? this.addFileExtension(value) : ""}`,
-        prevFilePath: `/${this.props.displayName}`
-      });
-    }
-
-    this.setState({ isDialogOpen: false });
-  };
-  render() {
-    const themeLogoLink = urljoin(this.props.appBase, this.props.baseDir);
-    const icon = this.getFileHandlerIcon();
-
-    return (
-      <React.Fragment>
-        <Nav>
-          <NavSection>
-            <a href={themeLogoLink} role="button" title="Home">
-              <ThemedLogo />
-            </a>
-            <div>
-              <H4 onClick={this.openDialog}>{this.props.displayName}</H4>
-              <EditableTitleOverlay
-                defaultValue={this.props.displayName}
-                isOpen={this.state.isDialogOpen}
-                onCancel={this.closeDialog}
-                onSave={this.confirmTitle}
-              />
-            </div>
-          </NavSection>
-          <NavSection>
-            <span className="icon">{icon}</span>
-            <LastSaved contentRef={this.props.contentRef} />
-          </NavSection>
-        </Nav>
-        <NotebookMenu contentRef={this.props.contentRef} />
-      </React.Fragment>
-    );
-  }
 }
 
 class Contents extends React.PureComponent<IContentsProps, IContentsState> {
@@ -222,12 +113,4 @@ const mapStateToProps = (
   };
 };
 
-const mapDispatchToProps = (dispatch: Dispatch) => ({
-  changeContentName: (payload: actions.ChangeContentName["payload"]) =>
-    dispatch(actions.changeContentName(payload))
-});
-
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps
-)(hot(module)(Contents));
+export default connect(mapStateToProps)(hot(module)(Contents));
